+++ src/Links.php
@@ -1,0 +1,143 @@
+<?php
+
+namespace Swis\JsonApi\Client;
+
+use ArrayAccess;
+use Illuminate\Contracts\Support\Arrayable;
+use Illuminate\Contracts\Support\Jsonable;
+use JsonSerializable;
+
+class Links implements ArrayAccess, Arrayable, Jsonable, JsonSerializable
+{
+    /**
+     * @var \Swis\JsonApi\Client\Link[]
+     */
+    protected $links = [];
+
+    /**
+     * @param \Swis\JsonApi\Client\Link[] $links
+     */
+    public function __construct(array $links)
+    {
+        $this->links = $links;
+    }
+
+    /**
+     * @param $key
+     *
+     * @return bool
+     */
+    public function __isset($key)
+    {
+        return $this->offsetExists($key);
+    }
+
+    /**
+     * @param $key
+     *
+     * @return mixed
+     */
+    public function __get($key)
+    {
+        return $this->offsetGet($key);
+    }
+
+    /**
+     * @param $key
+     */
+    public function __unset($key)
+    {
+        $this->offsetUnset($key);
+    }
+
+    /**
+     * @param $key
+     * @param $value
+     */
+    public function __set($key, $value)
+    {
+        $this->offsetSet($key, $value);
+    }
+
+    /**
+     * {@inheritdoc}
+     *
+     * @param mixed $offset
+     *
+     * @return bool
+     */
+    public function offsetExists($offset)
+    {
+        return isset($this->links[$offset]);
+    }
+
+    /**
+     * {@inheritdoc}
+     *
+     * @param mixed $offset
+     *
+     * @return mixed
+     */
+    public function offsetGet($offset)
+    {
+        return $this->links[$offset] ?? null;
+    }
+
+    /**
+     * {@inheritdoc}
+     *
+     * @param mixed $offset
+     * @param mixed $value
+     */
+    public function offsetSet($offset, $value)
+    {
+        $this->links[$offset] = $value;
+    }
+
+    /**
+     * {@inheritdoc}
+     *
+     * @param mixed $offset
+     */
+    public function offsetUnset($offset)
+    {
+        unset($this->links[$offset]);
+    }
+
+    /**
+     * {@inheritdoc}
+     *
+     * @return array
+     */
+    public function toArray()
+    {
+        return array_map(
+            function (Link $link) {
+                return $link->toArray();
+            },
+            $this->links
+        );
+    }
+
+    /**
+     * {@inheritdoc}
+     *
+     * @param int $options
+     *
+     * @return false|string
+     */
+    public function toJson($options = 0)
+    {
+        return json_encode($this->jsonSerialize(), $options);
+    }
+
+    /**
+     * {@inheritdoc}
+     *
+     * @return array|mixed
+     */
+    public function jsonSerialize()
+    {
+        return $this->toArray();
+    }
+}