+++ src/Traits/HasLinks.php
@@ -1,0 +1,33 @@
+<?php
+
+namespace Swis\JsonApi\Client\Traits;
+
+use Swis\JsonApi\Client\Links;
+
+trait HasLinks
+{
+    /**
+     * @var \Swis\JsonApi\Client\Links|null
+     */
+    protected $links;
+
+    /**
+     * @param \Swis\JsonApi\Client\Links|null $links
+     *
+     * @return $this
+     */
+    public function setLinks(Links $links = null)
+    {
+        $this->links = $links;
+
+        return $this;
+    }
+
+    /**
+     * @return \Swis\JsonApi\Client\Links|null
+     */
+    public function getLinks()
+    {
+        return $this->links;
+    }
+}